Rename Future::Policy and helpers.
"Future" and "new" policies are now just policies
following the removal of the old policies code.
