+++ app/helpers/policy_helper.rb
@@ -1,9 +1,9 @@
 module PolicyHelper
   def array_of_links_to_policies(policies)
-    policies.map { |policy| link_to policy.title, future_policy_path(policy), class: 'policy-link' }
+    policies.map { |policy| link_to policy.title, policy_path(policy), class: 'policy-link' }
   end
 
-  def future_policy_path(policy)
-    policy.is_a?(Future::Policy) ? policy.base_path : public_document_path(policy)
+  def policy_path(policy)
+    policy.base_path
   end
 end
