+++ app/models/classification.rb
@@ -154,7 +154,7 @@ class Classification < ActiveRecord::Base
   end
 
   def policies
-    Future::Policy.from_content_ids(policy_content_ids)
+    Policy.from_content_ids(policy_content_ids)
   end
 
 private
