+++ app/views/admin/classifications/edit.html.erb
@@ -21,7 +21,7 @@
         <p>This topic can't be deleted as it has associated policies:</p>
         <ul>
           <% @classification.policies.each do |policy| %>
-            <li><%= link_to(policy.title, future_policy_path(policy)) %></li>
+            <li><%= link_to(policy.title, policy_path(policy)) %></li>
           <% end %>
         </ul>
       </div>
