+++ test/support/document_controller_test_helpers.rb
@@ -158,7 +158,7 @@ module DocumentControllerTestHelpers
         refute_select "#related-policies"
       end
 
-      view_test "should render related future policies on #{document_type} pages" do
+      view_test "should render related policies on #{document_type} pages" do
         edition = create("published_#{document_type}", policy_content_ids: [policy_1["content_id"]])
         get :show, id: edition.document
         assert_select ".meta a", text: policy_1["title"]
