+++ test/unit/edition/related_policies_test.rb
@@ -80,10 +80,10 @@ class Edition::RelatedPoliciesTest < ActiveSupport::TestCase
   test "#policies returns future-policies, ignoring those that do not exist" do
     stub_content_register_policies
 
-    future_policy = Future::Policy.new(policy_1)
-    edition = create(:news_article, policy_content_ids: [future_policy.content_id, 'bogus-id'])
+    policy = Policy.new(policy_1)
+    edition = create(:news_article, policy_content_ids: [policy.content_id, 'bogus-id'])
 
     assert_equal 1, edition.policies.size
-    assert_equal future_policy.title, edition.policies[0].title
+    assert_equal policy.title, edition.policies[0].title
   end
 end
